+++ app/src/main/java/com/anytypeio/anytype/ui/page/PageFragment.kt
@@ -905,6 +905,7 @@ open class PageFragment :
         state.searchToolbar.apply {
             if (isVisible) {
                 searchToolbar.visible()
+                searchToolbar.focus(requireContext())
             } else {
                 searchToolbar.gone()
             }